feat: add new components for CSS Flexbox, Grid, Functions, and Responsive Design
- Introduced new components: CssFlexbox, CssGrid, CssFunctionsAndModernFeatures, ResponsiveDesign, and SemanticHtmlAndAccessibility.
- Updated HTMLCSSComponent to utilize the new components for better organization and modularity.
- Enhanced the structure and content of the HTML & CSS section for improved clarity and accessibility.
