+++ src/app/frontend/junior/html&css/components/index.ts
@@ -1,0 +1,5 @@
+export * from './css-flexbox';
+export * from './css-functions-and-modern-features';
+export * from './css-grid';
+export * from './responsive-design';
+export * from './semantic-html-and-accessibility';
